Some fixes and improvements.

+++ source/src/main/java/pl/edu/pw/elka/med/core/ItemSet.java
@@ -12,7 +12,7 @@ public interface ItemSet {
      *
      * @return wsparcie zbioru itemów
      */
-    double getSupport();
+    long getSupport();
 
     /**
      * Zwraca itemy ze zbioru.
